# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoundTrip Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/styles": {
      "post": {
        "tags": [
          "styles"
        ],
        "summary": "Create Style",
        "operationId": "create_style_api_v1_styles_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "styles"
        ],
        "summary": "List Styles",
        "operationId": "list_styles_api_v1_styles_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StyleRead"
                  },
                  "title": "Response List Styles Api V1 Styles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/styles/{style_id}": {
      "get": {
        "tags": [
          "styles"
        ],
        "summary": "Get Style",
        "operationId": "get_style_api_v1_styles__style_id__get",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Style Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "styles"
        ],
        "summary": "Update Style",
        "operationId": "update_style_api_v1_styles__style_id__put",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Style Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "styles"
        ],
        "summary": "Delete Style",
        "operationId": "delete_style_api_v1_styles__style_id__delete",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Style Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StyleCreate": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level",
            "default": 1
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "StyleCreate"
      },
      "StyleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "label",
          "level",
          "parent_id"
        ],
        "title": "StyleRead"
      },
      "StyleUpdate": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level",
            "default": 1
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "StyleUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}